FIX ALIGNMENT ISSUES BETWEEN sic_raster and Grid_poly
Going from a raster object to a SpatialGridDataFrame appears to shift the coordinates of the corner origins or something. This results in the SGDF and the derived SPDF mis-aligned from the original sic_raster. I removed the conversion to SGDF and used the rasterToPolygons() function. This also negated the need for setting a non-NA value.

Also minor clean-up of some original code relying on download.file() to get the tif from the LAN.